+++ src/greek_bess/__init__.py
@@ -1,3 +1,3 @@
 """Greek Battery Investment Stress Tester."""
 
-__version__ = "0.9.4"
+__version__ = "0.9.5"